# ms2/src/openapi/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lost & Found API",
    "description": "Backend API for the Lost & Found platform — college campus pilot with extensibility for apartments, offices, and transit operators.",
    "version": "1.0.0",
    "contact": { "name": "Lost & Found Team" }
  },
  "servers": [
    { "url": "http://localhost:4000/api/v1", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
      "Item": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "type": { "type": "string", "enum": ["LOST", "FOUND"] },
          "status": { "type": "string", "enum": ["PENDING_REVIEW", "ACTIVE", "CLAIMED", "RETURNED", "EXPIRED", "REMOVED"] },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "categoryId": { "type": "integer" },
          "locationName": { "type": "string", "nullable": true },
          "latitude": { "type": "number", "nullable": true },
          "longitude": { "type": "number", "nullable": true },
          "eventDate": { "type": "string", "format": "date-time" },
          "verificationQuestion": { "type": "string", "nullable": true, "description": "Only shown to authenticated non-owners" },
          "expiresAt": { "type": "string", "format": "date-time", "nullable": true },
          "returnedAt": { "type": "string", "format": "date-time", "nullable": true },
          "user": { "$ref": "#/components/schemas/PublicUser" },
          "images": { "type": "array", "items": { "$ref": "#/components/schemas/ItemImage" } },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "PublicUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "avatarUrl": { "type": "string", "nullable": true },
          "karmaPoints": { "type": "integer" }
        }
      },
      "ItemImage": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string" },
          "position": { "type": "integer" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/items": {
      "get": {
        "tags": ["Items"],
        "summary": "List active items with filters",
        "security": [],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["LOST", "FOUND"] } },
          { "name": "categoryId", "in": "query", "schema": { "type": "integer" } },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Full-text search" },
          { "name": "dateFrom", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "dateTo", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "lat", "in": "query", "schema": { "type": "number" } },
          { "name": "lng", "in": "query", "schema": { "type": "number" } },
          { "name": "radiusKm", "in": "query", "schema": { "type": "number" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["newest", "nearest"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Paginated list", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Item" } }, "meta": { "$ref": "#/components/schemas/PaginationMeta" } } } } } }
        }
      },
      "post": {
        "tags": ["Items"],
        "summary": "Create a LOST or FOUND item",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["type", "title", "description", "categoryId", "eventDate"], "properties": { "type": { "type": "string", "enum": ["LOST", "FOUND"] }, "title": { "type": "string" }, "description": { "type": "string" }, "categoryId": { "type": "integer" }, "eventDate": { "type": "string", "format": "date-time" }, "locationName": { "type": "string" }, "latitude": { "type": "number" }, "longitude": { "type": "number" }, "imageUrls": { "type": "array", "items": { "type": "string" }, "maxItems": 5 }, "verificationQuestion": { "type": "string", "description": "Required for FOUND items" }, "verificationAnswer": { "type": "string", "description": "Required for FOUND items — never returned in responses" }, "handoverPointId": { "type": "string", "format": "uuid" } } } } }
        },
        "responses": {
          "201": { "description": "Item created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "tags": ["Items"],
        "summary": "Get item by ID",
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Item detail" }, "404": { "description": "Not found" } }
      },
      "patch": {
        "tags": ["Items"],
        "summary": "Update item (owner or admin)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Updated item" }, "403": { "description": "Forbidden" } }
      },
      "delete": {
        "tags": ["Items"],
        "summary": "Soft delete item",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
...
    "/items/{id}/matches": {
      "get": {
        "tags": ["Matches"],
        "summary": "Get potential matches for an item (owner only)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Matches list" } }
      }
    },
...
    "/matches/{id}/dismiss": {
      "post": {
        "tags": ["Matches"],
        "summary": "Dismiss a match",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Match dismissed" } }
      }
    },
...
    "/admin/reports/{id}/resolve": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Resolve a report",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Report resolved" } }
      }
    },
    "/admin/items": {
      "get": {
        "tags": ["Admin"],
        "summary": "List items by status (staff only)",
        "parameters": [{ "name": "status", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Items" } }
      }
    },
    "/admin/items/{id}/moderate": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Approve or reject an item",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["action"], "properties": { "action": { "type": "string", "enum": ["approve", "reject"] }, "reason": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Moderated" } }
      }
    },
    "/admin/users/{id}/ban": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Ban a user (admin only)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "User banned" } }
      }
    },
    "/admin/users/{id}/unban": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Unban a user (admin only)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "User unbanned" } }
      }
    },
...
  }
}